feat: design and page refactoring
