+++ app/admin/settings/page.tsx
@@ -1,0 +1,9 @@
+"use client";
+
+import { UnderDevelopmentAlert } from "@/app/components/UnderDevelopmentAlert";
+
+const Settings = () => {
+  return <UnderDevelopmentAlert />;
+};
+
+export default Settings;
