+++ app/bottomNavbar.tsx
@@ -1,7 +1,7 @@
 // when using reacthook or react components, use client
 "use client";
 import React from "react";
-import { useSearchParams } from "next/navigation";
+import { usePathname, useSearchParams } from "next/navigation";
 
 import { useSession } from "next-auth/react";
 import {
@@ -12,41 +12,67 @@ import {
 } from "react-icons/fa6";
 import { FaHistory } from "react-icons/fa";
 import clsx from "clsx";
-import { twMerge } from "tailwind-merge";
 
 const BottomNavbar = () => {
   const { status, data: session } = useSession();
 
   const searchParams = useSearchParams();
 
+  const pathname = usePathname();
+
   return (
     <>
       {searchParams.get("tab") !== "admin" && (
         <div className="btm-nav">
-          <button className="text-blue-400">
+          <a
+            href="/member/borrow"
+            className={clsx("text-gray-500", {
+              active: pathname === "/member/borrow",
+            })}
+          >
             <FaBookOpenReader />
             <span className="btm-nav-label">本を借りる</span>
-          </button>
-          <button className="active text-pink-400">
+          </a>
+          <a
+            href="/member/return"
+            className={clsx("text-gray-500", {
+              active: pathname === "/member/return",
+            })}
+          >
             <FaArrowRotateLeft />
             <span className="btm-nav-label">本を返す</span>
-          </button>
-          <button className="text-orange-400">
+          </a>
+          <a
+            href="/member/history"
+            className={clsx("text-gray-500", {
+              active: pathname === "/member/history",
+            })}
+          >
             <FaHistory />
             <span className="btm-nav-label">履歴</span>
-          </button>
+          </a>
         </div>
       )}
       {searchParams.get("tab") === "admin" && (
         <div className="btm-nav">
-          <button className="text-blue-400">
+          <a
+            href="/admin/books?tab=admin"
+            className={clsx("text-gray-500", {
+              active: pathname === "/admin/books",
+            })}
+          >
             <FaBook />
-            <span className="btm-nav-label">本を登録</span>
-          </button>
-          <button className="active text-pink-400">
+            <span className="btm-nav-label">本の管理</span>
+          </a>
+          <a
+            href="/admin/settings?tab=admin"
+            className={clsx("text-gray-500", {
+              active: pathname === "/admin/settings",
+            })}
+          >
             <FaWrench />
             <span className="btm-nav-label">設定</span>
-          </button>
+          </a>
         </div>
       )}
     </>
