+++ app/components/underDevelopmentAlert.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+import { FaExclamationCircle } from "react-icons/fa";
+
+export const UnderDevelopmentAlert = () => {
+  return (
+    <div role="alert" className="flex justify-center alert my-2 text-gray-600">
+      <FaExclamationCircle size={26} />
+      <span>開発中</span>
+    </div>
+  );
+};
