+++ app/member/history/page.tsx
@@ -1,0 +1,9 @@
+"use client";
+
+import { UnderDevelopmentAlert } from "@/app/components/UnderDevelopmentAlert";
+
+const History = () => {
+  return <UnderDevelopmentAlert />;
+};
+
+export default History;
