+++ app/member/return/page.tsx
@@ -1,0 +1,9 @@
+"use client";
+
+import { UnderDevelopmentAlert } from "@/app/components/UnderDevelopmentAlert";
+
+const Return = () => {
+  return <UnderDevelopmentAlert />;
+};
+
+export default Return;
