+++ tsconfig.json
@@ -22,6 +22,13 @@
       "@/*": ["./*"]
     }
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts", "app/rentReturn/page.tsx", "app/apitest/page.jsx"],
+  "include": [
+    "next-env.d.ts",
+    "**/*.ts",
+    "**/*.tsx",
+    ".next/types/**/*.ts",
+    "app/member/borrow/page.tsx",
+    "app/apitest/page.jsx"
+  ],
   "exclude": ["node_modules"]
 }
